separated the functionality from the controller and moved it into a new file, imported the new file into client/index.html
