+++ client/index.html
@@ -6,7 +6,9 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="icon" type="image/x-icon" href="./favicon.ico">
   <link rel="stylesheet" href="style.css">
+  <script src="./utils.js"></script>
   <script defer src="./app.js"></script>
+  
   <title>FullStack</title>
 </head>
 <body>
@@ -34,7 +36,7 @@
 
   <footer class="footer">
     <h3 class="notes">Be sure to follow the tutorial through the steps as per the instructions in README.md!</h3>
-    <h5>And do not forget to import style.css and app.js in the head of this html file!</h5>
+    <h5>And do not forget to import style.css, app.js and utils.js in the head of this html file!</h5>
   </footer>
 </body>
 </html>
